+++ perf/ReaderTask/traverseSeqArrayWithIndex.ts
@@ -1,11 +1,11 @@
 import * as Benchmark from 'benchmark'
 import * as A from '../../src/Array'
-import * as _ from '../../src/Task'
+import * as _ from '../../src/ReaderTask'
 import { pipe } from '../../src/function'
 
 /*
-A.traverseWithIndex(_.ApplicativeSeq) x 310 ops/sec ±10.89% (31 runs sampled)
-_.traverseSeqArrayWithIndex x 1,324 ops/sec ±6.92% (20 runs sampled)
+A.traverseWithIndex(_.ApplicativeSeq) x 247 ops/sec ±4.98% (36 runs sampled)
+_.traverseSeqArrayWithIndex x 1,901 ops/sec ±10.34% (66 runs sampled)
 Fastest is _.traverseSeqArrayWithIndex
 */
 
@@ -18,13 +18,13 @@ suite
     return pipe(
       as,
       A.traverseWithIndex(_.ApplicativeSeq)((_i, a) => _.of(a))
-    )()
+    )(undefined)()
   })
   .add('_.traverseSeqArrayWithIndex', function () {
     return pipe(
       as,
       _.traverseSeqArrayWithIndex((_i, a) => _.of(a))
-    )()
+    )(undefined)()
   })
   .on('cycle', function (event: any) {
     // tslint:disable-next-line: no-console