+++ perf/StateReaderTaskEither/sequenceArray.ts
@@ -1,11 +1,11 @@
 import * as Benchmark from 'benchmark'
 import * as A from '../../src/Array'
-import * as _ from '../../src/Task'
+import * as _ from '../../src/StateReaderTaskEither'
 import { pipe } from '../../src/function'
 
 /*
-A.sequence(_.ApplicativePar) x 390 ops/sec ±13.15% (44 runs sampled)
-_.sequenceArray x 3,460 ops/sec ±7.30% (57 runs sampled)
+A.sequence(_.Applicative) x 5,930 ops/sec ±0.39% (90 runs sampled)
+_.sequenceArray x 21,057,757 ops/sec ±0.84% (88 runs sampled)
 Fastest is _.sequenceArray
 */
 
@@ -14,11 +14,11 @@ const suite = new Benchmark.Suite()
 const as = pipe(A.range(0, 1000), A.map(_.of))
 
 suite
-  .add('A.sequence(_.ApplicativePar)', function () {
-    return pipe(as, A.sequence(_.ApplicativePar))()
+  .add('A.sequence(_.Applicative)', function () {
+    pipe(as, A.sequence(_.Applicative))
   })
   .add('_.sequenceArray', function () {
-    return pipe(as, _.sequenceArray)()
+    pipe(as, _.sequenceArray)
   })
   .on('cycle', function (event: any) {
     // tslint:disable-next-line: no-console